Exercise 4.21

+++ part4/blog/controllers/blogs.js
@@ -28,6 +28,15 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+  if ( blog.user.toString() !== decodedToken.id.toString() ){
+    return response.status(401).json({ error: 'only the user that created a blog can delete it' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
